Changes for DERBY-689, contributed by Rick Hillegas (rick.hillegas@sun.com)
Rick says:

Various improvements to compatibility test:

1) Moved all JUnit tests under a ...functionTests/tests/junitTests as Myrna and John requested.

2) Made the test harness sanity check the classpath as David requested.

3) Abstracted a CompatibilitySuite out of the original test to make it easier to add more test classes, as Francois requested.

4) Added combinations for the 10.1.2.0 client and server

5) The derbyall diff tool was enhanced to fabricate an empty canon file if one doesn't exist in the codeline. This eliminates the need to create empty canons for JUnit tests.

===

David says:

I made a slight modification -- I moved the CompatibilityTest to 
derbynetclientmats instead of derbynetmats, as it's intended only for
the network client, not for the DB2 driver

I'd also like to clarify that the "empty canon file" is not created on
disk -- it's only created as an InputStream within the test harness.  Nice
fix, no more empty files to create for tests that have no output.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@332090 13f79535-47bb-0310-9956-ffa450edef68