DERBY-239, committing next phase checkin on behalf of Suresh Thalamati.
This patch adds code to support real-time online backup with unlogged
operations. A consistent backup can not be made if there are pending
transactions with unlogged operations or if unlogged operations 
occur when backup is in progress. Because container files can  be copied 
to the backup before the transaction is committed and the data pages are 
flushed as part of the commit. As there is no transaction log for unlogged
operations, while restoring from the backup database can not be restored to 
a consistent state. To make a consistent online backup in this scenario, this patch:

1)  blocks  online backup until all the transactions with unlogged operation are
committed/aborted. 
2)  implicitly converts all unlogged operations to logged mode for the duration
of the online backup, if they are started when backup is in progress. 

This patch also adds a test to test the online backup in parallel with some DML,

> DDL and unlogged operations. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@345355 13f79535-47bb-0310-9956-ffa450edef68