merging fix for DERBY-715 from trunk to 10.1 branch.
Sometimes (timing dependent) a lock deadlock was incorrectly returned to the
user as a lock timeout.  The code in LockSet was using the wrong variable
to determine if a deadlock had been detected, that variable sometimes was
right and sometimes wrong.  Changed all the code to use the direct return
from the deadlock detection routine.  Added a test that causes 5 deadlocks,
previous to the fix this test would always report at least 1 timeout on
my single processor, windows xp, laptop.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@353812 13f79535-47bb-0310-9956-ffa450edef68