DERBY-239, fixing onlinebackuptest1.
committing on behalf of Suresh Thalamati.

This pacth  fixes store/onlineBackupTest1.java failure on non-windows envirorment.   Problem  was  unlogged operations thread and insert thread are working on the  same connection. Test was failing becuase  insert thread  was committing the unlogged operation that was suppose to block the backup. 
This pacth modified the test , so that  these threads works on different conenctions, 
 

 Test passed on Windows XP and Linuix. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@354523 13f79535-47bb-0310-9956-ffa450edef68