committing onlinebackup_4.diff for DERBY-239.
committing on behalf of Suresh Thalamati.

This patch makes online backup call to wait/fail when unlogged operations like
create index are pending. It also fixes derby-523 bug my making the existing
log archive backup procedure to wait for the unlogged operation to complete. 

-- Two new procedures are added to allow the users to make backup wait/fail
when unlogged operations are pending. 
-- prevents users starting backup in an non-idle transactions to avoid backup
blocking forever if users starts backup in the same transaction as an
unlogged operation. 
-- backup is not really transactional , to avoid any locking issues in the
futures; backup procedures ends the transaction by implicitly doing commit
when it is successful or rollback on any errors. 

A new backup test is added to store suite to test the above scenarios. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@354898 13f79535-47bb-0310-9956-ffa450edef68