DERBY-573: Add support for optimizer hints, provided by the user as SQL comments. Also add upgrade infrastructure for 10.2 release and support new optimizer mechanism to work correctly under soft-upgrade for databases at 10.1 level.
I have the patch for optimizer overrides support in Derby. Alongwith the patch, I have attached the updated functional spec to the JIRA entry DERBY-573.

Majority of the changes went into the sqlgrammar.jj because Derby engine already has support for them internally. It is the parser that needs to recognize these overrides and pass it on to through the query nodes. The parser now looks for character sequence -- DERBY-PROPERTIES (case insensitive and space between -- and D is optional) and once it finds that, it looks for propertyName=value pairs on that same comment line in parser's propertyList method. The parser does the basic check to make sure that the same property is not used more than once for a given table. The remaining checks on the properties like checking the existence of user specified index etc are done in the bind phase.

I also changed the metadata.properties file to use --DERBY-PROPERTIES rather than old PROPERTIES clause to supply optimizer overrides. In addition, added \n at the end of the optimier override comment lines to make sure the comment line does not get concatenated with the next line of the sql.

Import.java had to be changed to user --DERBY-PROPERTIES rather than PROPERTIES.

Added a new test optimizerOverrides.sql which runs in both embedded and network server mode.

Rerunning all the tests after syncing the codeline to make sure nothing has broken. An earlier run of the tests before the sync came out clean.

I plan to next work on exposing these overrides through runtime statistics so that user can verify that the optimizer overrides are getting used.

Submitted by Mamta Satoor (msatoor@gmail.com)



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@356562 13f79535-47bb-0310-9956-ffa450edef68