DERBY-239, committing on behalf of Suresh Thalamati.
This patch adds code to support online backup when  jar operations 
are running parallel to the backup. Jar files are not logged, but the 
system catalogs updates are logged when a jar file is added/replaced. 
If the jar file operations are allowed during the backup, system catalog 
(sys.sysfiles) table in the backup database can have a reference to a jar file 
that does not exist in the backup database. And also backup can contain 
partial written jar files. To make a consistent online backup, this patch:

1) Makes Backup operation wait/fail for all the jar operations activity in progress to complete.
2) Blocks jar file operations when a backup is in progress.

This patch also adds a new test to test the online backup 
with jar operations.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@357269 13f79535-47bb-0310-9956-ffa450edef68