DERBY-750, committing on behalf of Suresh Thalamati.
By looking at the stacks,  my suspicion is backup is looking at a table that 
is not completely created yet, container file might be there but the  
first page (Alloc page) might have been created.  

It might take some time for me to fix this one.  As a temporary solution,  
attached patch comments out the section of the test that might be hitting this 
problem. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@357272 13f79535-47bb-0310-9956-ffa450edef68