DERBY-573: Enhance RUNSTAT output to show user specified optimizer hints are bing used.
Here is the checkin message from the contributor:

I have attached a patch named Derby573OptimizerOverridesAndRunTimeStatistics011206.txt to JIRA Derby573 Provide support for optimizer overrides in Derby. This patch enables users to see the optimizer overrides specified in the sql as part of runtime statistics info. This is achieved by changing the 
generator so that these properties get passed from compile time to execute time. This change in generate phase can be found in FromBaseTable,
BaseJoinStrategy and JoinNode. The changes in the other classes is for returning the correct number of arguments to the scan. That change is in 
getScanArgs method. 

In addition, I have changed the existing lang/optimizerOverrides.sql to test this patch. derbyall suite has run fine on my Windows XP m/c with Sun's jdk14.

Submitted by Mamta Satoor (msatoor@google.com)


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@369619 13f79535-47bb-0310-9956-ffa450edef68