DERBY-756: Close scan controller in DistinctScalarAggregateResultSet to avoid memory leak.
Submitted by Manish Khettry (manish_khettry@yahoo.com)

Here is more info from the contributor:
It looks like the scan controller created in DistinctScalarAggregateResultSet is never closed. This causes the scanController instance variable in RAMTransaction to continue to grow with each open/close. 

Shouldn't DistinctScalarAggregateRS implement a close or a finish method (not sure what the difference is) and close the scan controller there.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@370978 13f79535-47bb-0310-9956-ffa450edef68