Fix for DERBY-757, ported from trunk to 10.1 branch.
Now that this routine can be called while concurrently with XA transactions
being created and committed, checks have to be made to make sure the object
being searched for does not go null while we are looking at it.  The search
is for existing xacts, so skipping when we find a null is fine.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@371085 13f79535-47bb-0310-9956-ffa450edef68