DERBY-599 committing on behalf of: Sunitha Kambhampati
Problem:
setBlob(i,blob) does not set the length of the stream in the blob and but instead passes a -1 for stream length. 
During the normalization process, setBlob.normalize(DTD,DVD) calls SQLBlob.setWith. 
The setWidth is called in order to compare the length of the blob value to the maximum width of the column and to throw a truncation error in case the value wont fit into the column.  setWidth() calls SQLBinary.getLength().  If tvalue is a stream, the getLength() method checks for streamLength value and if the streamLength value is set to -1 (ie unknown) , it calls getBytes().length() which calls getValue() and this is where the entire stream is getting materialized.  

This patch fixes DERBY-599 so using setBlob call, will not  materialize the entire blob into memory. 
- changes to setBlob to pass the length of the blob value instead of -1.  The length of the blob value passed into setBlob can be obtained by calling Blob.length() which returns a long.
- move the -ve length check from setBinaryStream to setBinaryStreamInternal since setBlob will not pass a -1 for length.
- change setBinaryStreamInternal to take the length parameter as a long instead of int.  
- Currently Derby allows max value of 2G-1 ( Max value of an int) for blobs. Add check to ensure that if a stream with a length > max value of int is passed, an error is thrown. Use an existing error message 
'The resulting value is outside the range for the data type {0}'



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@373356 13f79535-47bb-0310-9956-ffa450edef68