Derby-239, onlinebackup_8.diff patch committed on behalf of Suresh Thalamati
This patch addresses the improvements suggested by Oystein in his review of the
first online backup patch(onlinebackup_1.diff) and also resolved some more
online backup issues.

-- fixed comments and moved a duplicate code in write page and the backup of
container into a separate method.

fixes with this patch:

-- backup of container was using the same encryption buffer as container
read/writes, this requires backup of the container and read/writes are
synchronized to avoid corrupting the encryption buffer. This patch modified
the backup of container code to use it's own temporary encryption
buffer to avoid the buffer corruption.

-- added code to prevent truncation of log during checkpoint and disabling of
log archival does not delete log files that are yet to be copied
into the backup, if backup is running in parallel.

-- In-place compress is blocked during backup and vice-versa, until the
compress nested transaction commits. This change is needed because compress
does a special checkpoint to avoid redo on the truncated pages. If backup is
running in parallel, the checkpoint backup is based on can be earlier than
checkpoint done by the truncate operation. Without the blocking during
restore from backup, recovery will fail if it needs redo ant log records on
the truncated pages. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@373380 13f79535-47bb-0310-9956-ffa450edef68