DERBY-861
issue is that  just prior to this query the test deletes and commits some rows.  In one phase of this locking test
the table is configured to have one row per page.  In this case as soon as the delete commits, the post commit background
thread can be scheduled to reclaim the deleted rows.  This scheduling is completely at the discretion of a number of factors including: OS/JVM/processor availbility .  To get reproducible results for this query I added a wait after the delete.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@375818 13f79535-47bb-0310-9956-ffa450edef68