DERBY-210 cleanup patch in preparation for change Attaching a patch 'derby-210-patch1.diff' which removes the list 'RollbackOnlyListeners_' from org.apache.derby.client.am.Connection class since this is not being used. This is an incremental patch and does not solve the issue fully.
While working on this issue, I found one of the causes of memory leaks is object references stored in the lists in the Connection class. On looking at the lists, I found that 'RollbackOnlyListeners' is not being used anywhere in the client code. Submitting this patch to remove this. 

Contributed by Deepa Remesh



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@376874 13f79535-47bb-0310-9956-ffa450edef68