DERBY-614 patch 2 - clear splitQRYTDA field for reuse of the resultset Bryan said ..
 With my first patch to this bug, I introduced a problem in this processing; the
splitQRYDTA field was not cleared when the statement was closed. This meant that
if that splitQRYDTA field just happened to have some pending query data from the
previous statement at the point where it was re-used, the pending query data would
erroneously stick around and be returned as the first result for the new statement.

Contributed by Bryan Pendleton



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@378552 13f79535-47bb-0310-9956-ffa450edef68