DERBY-1010 - setTransactionIsolation can cause ava.sql.SQLException: Invalid operation: statement closed in some contexts with xa
Client has a Statement object that it reuses to send SET CURRENT ISOLATION commands to the server.  Added a check to make sure this statement was not closed before trying to reuse it.

This change takes the checkDataSource test a bit further (DERBY-435) but there seem to be issues with isolation being set correctly with xa_join.  I will investigate that and file a Jira issue



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@379993 13f79535-47bb-0310-9956-ffa450edef68