DERBY-788: 'store/encryptionKey.sql' fails on Solaris 10
Contributed by Kristian Waagan <Kristian.Waagan@Sun.COM>.

'DERBY-788-1a.diff' is a patch that replaces the 16 byte key with a 8
byte key for the DES encryption algorithm used in the test
'store/encryptionKey.sql'. The test still exercises the intended parts
of the Derby code (according to my own analysis and that of
Sunitha). The patch makes the test stop failing on Solaris10 (with
security provider 'SunPCKS11-Solaris') .

One test case is changed; where it previously tested a key that was of
incorrect length (longer than 8 bytes, shorter than 16), it now tests
an incorrect key of the correct length. It is not possible to make the
test pass on Solaris10 with a key of incorrect length, because keys of
length less than 8 bytes are caught by a check in the code, and keys
longer than 8 bytes cause the underlying security provider to fail
('SunPCKS11-Solaris'). This issue will be addressed in DERBY-1001
(rewrite of test to a JUnit-test).


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@381333 13f79535-47bb-0310-9956-ffa450edef68