DERBY-805: Apply Part I of the fix. Refer to JIRA for complete writeup on all the phases. This Phase broadly addresses:
When remembering "truly the best" access path for an Optimizable, we have to 
keep track of which OptimizerImpl the "truly the best" access is for.  
In most queries there will only be one OptimizerImpl in question, but in 
cases where there are nested subqueries, there will be one OptimizerImpl
for every level of nesting, and each OptimizerImpl might have its own
idea of what this Optimizable's "truly the best path" access path really is.
So whenever we save a "truly the best" path, we take note of which Optimizer
told us to do so.  Then as each level of subquery finishes optimization,
the corresponding OptimizerImpl can load its preferred access path into this
Optimizable's trulyTheBestAccessPath field and pass it up the tree, until 
eventually the outer-most OptimizerImpl can choose to either use the best path
that it received from below (by calling "rememberAsBest()") or else use the
path that it found to be "best" for itself.

Submitted by Army Brown (qozinx@sbcglobal.net)

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@381859 13f79535-47bb-0310-9956-ffa450edef68