DERBY-709, committing on behalf of Suresh Thalamati
This patch (derby-709_latest.diff) fixes the problems appreared with earlier
patch when tests are run againest JARS.

--- correctly catch the Security Exception when there is no permission to get
a canonical path.
--- Modified the deryb_tests.plocy to allow deltes in
${user.dir}${/}extinout${/}- .,
It is needed because, backup tests deletes the old copy of the backup at the
backup location.
--- set noSecurityManager = true for the tests that can not be run under
security manager due to bug : 1066
--- changes st_1 properties to use external tests dirs , so that tests does not
required "user.dir" permission to create parent directories for the backup
paths.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@381883 13f79535-47bb-0310-9956-ffa450edef68