DERBY-1073: Reset optimizer timeout for subqueries after new scoped predicates are pushed into it for next round of evaluation.
If we have predicates that were pushed down to this OptimizerImpl from an
outer query, then we reset the timeout state to prepare for the next round
of optimization.  Otherwise if we timed out during a previous round and then
we get here for another round, we'll immediately "timeout" again before
optimizing any of the Optimizables in our list.  This is okay if we don't
have any predicates from outer queries because in that case the plan we find
this round will be the same one we found in the previous round, in which
case there's no point in resetting the timeout state and doing the work a
second time.  But if we have predicates from an outer query, those predicates
could help us find a much better plan this round than we did in previous
rounds--so we reset the timeout state to ensure that we have a chance to
consider plans that can take advantage of the pushed predicates.

Submitted by Army Brown (qozinx@sbcglobal.net)


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@382953 13f79535-47bb-0310-9956-ffa450edef68