DERBY-670 - improve automatic space reclamation of long rows and long columns.
This patch provides a non-upgrade solution to the problem.  It uses existing
row and column format information to determine at delete time if the row or
column is "long".  If so it schedules a post commit reclamation immediately
rather than waiting for all rows on a page to be deleted.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@383663 13f79535-47bb-0310-9956-ffa450edef68