DERBY-937
Fixing typo in the no timeout setting for this test.  it should have
been derby.optimizer.noTimeout=true not derby.optimize.noTimeout=true.
Unfortunately did this did not fix the diffs I am seeing on my machine, but
it clearly makes sense for this test to set no timeout to eliminate the
variance of running this test on a number of different speeds of hardware.
Maybe this change will help others, at least it will eliminate one variable
as we continue to track down the cause.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@384083 13f79535-47bb-0310-9956-ffa450edef68