DERBY-1080: Connection reset when using security mechanism=EUSRIDPWD results in protocol error
Patch contributed by Sunitha Kambhampati

For eusridpwd case, the client sends the encrypted userid and password
sectkns as part of SECCHK. The protocol error was happening because we
only read the 2 sectkns if the database.decryptedUserId ,
database.decryptedPassword is null, otherwise we think we have already
read this. Thus on a connection reset,if the decryptedUserId and
decryptedPassword are not reset, server assumes we have seen more SECTKN's
and thus we throw error too many codepoints.

Patch adds
-- code to reset the security mechanism related variables on a connection re-use
-- regression test to testSecMec.java 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@385857 13f79535-47bb-0310-9956-ffa450edef68