DERBY-668: SysInfo does not print the right info in class loader scenarios
This change modifies the sysinfo tool to print additional information about
the classes and jars that it examines. For each class that sysinfo examines,
it now calls getProtectionDomain().getCodeSource() to retrieve information
about the source of that file, which might be: an entry in the application
classpath, an entry in a class loader location list, a jar fetched due to
being listed in the manifest entry of another jar, a standard extension
in the JRE's extensions directory, a jar installed into the application
server, or any of various other possibilities.

The additional information about the class's origin is displayed in the
sysinfo output.

Note that when sysinfo runs under a Java security manager, it may need
special permissions to access this additional information, including the
permission to read the java.class.path property, and the permission to
call getProtectionDomain on a class. If sysinfo is not granted these
permissions, it will display an error message about the security problem
in place of displaying the class origin information.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@387599 13f79535-47bb-0310-9956-ffa450edef68